remove helper method of ces
