+++ opentelekomcloud/resource_opentelekomcloud_ces_alarmrule.go
@@ -505,7 +505,7 @@ func resourceAlarmRuleDelete(d *schema.ResourceData, meta interface{}) error {
 		return nil
 	})
 	if err != nil {
-		if isCESResourceNotFound(err) {
+		if isResourceNotFound(err) {
 			log.Printf("[INFO] deleting an unavailable %s: %s", nameCESAR, arId)
 			return nil
 		}
